services/ingestion/ocr_document/requirements.txt
boto3>=1.26.0
Pillow>=10.4.0
# Updated for Python 3.13 compatibility
